Updated implmentation of TinyLeaders.java
Made several changes. The commander card should now be in the main board. It is found and verified against the deck name attribute. Both main board and sideboard are checked for the proper number of cards (50 and 10 respectively). Main board and sideboard card colors are also verified against the colors of the commander. I haven't written the server plugin for the game type yet, so this isn't tested. It should be very close to being right, though.