SQL query optimisation
There will be always 0 or 1 distinct card. If we don't limit this query it will find a card, and still will go through the whole table, which is unneccesary and have performance impact.
Also changed landTypes declaration from new TreeSet() to  new TreeSet<>() to avoid "unchecked cast" warning in IDE.