Forge of Heroes can place counters only on creatures or planeswalkers.
If a Commander becomes other type after the activation but before the resolution, the ability will be countered by the game, due to invalid target. For instance: player A plays their commander and it resolves. Then they activate Forge of Heroes' second ability. In response player B flashes Song of the Dryads on that commander. Song resolves, commander becomes a Forest and loses all it's other types. Forge of Heroes is countered by the game since land is not an allowed type.