Fix Soul Strings (Still an error)
The code for the X value was not working, it didn't get the X value from being cast. I have set it up so it can be done correct, but I'm not sure how to finish it. This is why line 59 is empty. Hopefully someone can see what needs to be done.