6079 - Fixing conditional damage order, conditions
As of c482fad, a number of spells had their conditional damage
reworded in oracle text. As a part of these changes, some of the
spells had their order of effects inverted, with the damage
after the condition was happening before the condition was met.
In addition, Brimstone volley was changed from a Morbid condition
to a Hellbent condition. This commit corrects those typos.