Fixed issues with maddening hex and made curse attachments more clear
Fixed issue with maddening hex where it wasn't detaching from players when the ability triggered, and instead just adding itself to a growing list of attached players. This caused the attachment arrows to show increasingly more opponents that it was attached to, until eventually all opponents were pointed at and the arrows were useless. As well, when this happened, the curse would fail to move to a different opponent, causing the same opponent to continue to take damage.

Also added a log message when curses are reattached to a different player (either with Maddening Hex or Lynde, Cheerful Tormentor) that shows the new player that it is attached to, to help prevent confusion in case the arrows break later.

Finally, made it so that all permanents attached to things that aren't also permnents show what they are attached to in their tooltip to also help avoid this confusion in the future