WIP: implement squad
A few tests are wrong at the moment, on different way of cloning/copying a spell cast with an additional squad.
test_CopyingETBTriggerMustKeepSquadStatus : Sometimes the lithoform engine seems to not be activated at the right moment. Not sure how to fix that.
test_CloneMustNotCopySquad: cloning a squad permanent should never produce more tokens, but as of now the clone comes with 2 additional tokens. Adding a source id in SquadAbility.getActivationKey seems to help with this one, but it feels incorrect.
test_CopyingSpellMustKeepSquadStatus: there is an additional token produced. I suspect addSubAbility to not work like I was expecting.
Latest two might also be not understanding how instance work in details. Should I copy instances for each usage?