Created EntersBattlefieldCastTriggeredAbility
An extension to EntersBattlefieldAllTriggeredAbility, it further restricts the trigger based on whether or not the permanent was cast.

Modified CreatureWasCastWatcher -> PermanentWasCastWatcher to watch for all permanents. This should have no effect on cards which already use it, since it's just watching for more things. This is used by EntersBattlefieldCastTriggeredAbility.

Made generateTriggerPhrase in EntersBattlefieldAllTriggeredAbility protected to facilitate code reuse in abilities which only need to slightly modify the trigger text.