[MOM] fix Tribute of the World Tree
My understanding of the issue was that getValue("permanentEnteringBattlefield")
needed to be contextualized for the game when  the apply does alter the game state.
Calling game.getPermanent() with the entering permanent uuid seems to be working.

I do not understand fully how apply was doing both effects at once, it seemed like
there was only a single apply call anyway. If anyone got more insight on that, I am
interested.