Hopefully code should be complete
Now that I understand collation generation better, I was able to complete this code.
I am still expecting it to fail because mutate creatures have been omitted from beta (but are included in collation)

Because I have a preference for Japanese Collation over US, I included commented out code for Japanese collation as an alternative.

For simplicity, I have used the base prints for rares and mythics only. So that is one further possible improvement.